requirements.txt
google-generativeai
instagrapi
google_images_search
requests
dotenv